% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtracker.R
\name{covidtracker_date_range}
\alias{covidtracker_date_range}
\alias{covidtracker_actions}
\alias{covidtracker_actions_range}
\alias{covidtracker_actions_dates}
\title{Access COVID Tracker API}
\usage{
covidtracker_date_range(from, to)

covidtracker_actions(country, date)

covidtracker_actions_range(countries, from, to, verbose = TRUE)

covidtracker_actions_dates(countries, dates, verbose = TRUE)
}
\arguments{
\item{from}{the first date to return data for}

\item{to}{the last date to return data for}

\item{country}{a ISO 3166-1 alpha-3 country code}

\item{date}{the date to return data for}

\item{countries}{a vector of ISO 3166-1 alpha-3 country codes}

\item{verbose}{show progress of data download}

\item{dates}{multiple dates to return data for}
}
\description{
Access data on government reactions to COVID-19 pandemic
from the COVID Tracker API.
}
\details{
[covidtracker_date_range()] return stringency index for all countries for the given date range.
[covidtracker_actions()] return the governmental actions for the given date and country.
[covidtracker_actions_range()] wrapper functions to return government actions for multiple countries and date range.
}
\examples{
cta <- covidtracker_actions("FIN", "2020-03-15")


}
\references{
\url{https://www.bsg.ox.ac.uk/research/research-projects/coronavirus-government-response-tracker}
}
