% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidtracker.R
\name{data_frame_actions}
\alias{data_frame_actions}
\alias{data_frame_stringency}
\title{Extract data from covidtracker objects}
\usage{
data_frame_actions(x, ...)

data_frame_stringency(x)
}
\arguments{
\item{x}{a [covidtracker_actions] object to extract data from}

\item{...}{further arguments to methods.}
}
\description{
Extract data from covidtracker objects
}
\details{
[data_frame_actions] extract policy actions together with date and country code
[data_frame_stringency] extract cases, deaths and stringency index together with date and country code
}
