% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_csv.R
\name{covidtracker_update_actions_csv}
\alias{covidtracker_update_actions_csv}
\alias{covidtracker_create_actions_csv}
\title{Create and store API data as CSV files}
\usage{
covidtracker_update_actions_csv(
  file,
  from = NULL,
  to = Sys.Date(),
  verbose = TRUE
)

covidtracker_create_actions_csv(file, countries, from, to, verbose = TRUE)
}
\arguments{
\item{file}{csv-file to store or update}

\item{from}{the first date to return data for}

\item{to}{the last date to return data for}

\item{verbose}{show progress of data download}

\item{countries}{a vector of ISO 3166-1 alpha-3 country codes}
}
\description{
Create and store API data as CSV files
}
\details{
[covidtracker_create_actions_csv()] create a csv file to store response information in
[covidtracker_create_update_csv()] update an existing csv file with new data
}
